refactor(core): runtime/reader.ts owns the ambient reader
The null-frame stack, get, the guarded readable behind h.reader,
readSource (Atom/AtomRef/plain through a registry get), and
guardRegistryRead (throw inside a registry read -> Effect.die) move to
one module; h.ts and On become short callers. No behavior change.
